fix(gates): collision-proof benchmark result path and shared gate (#480)
